Use defined embeds domain to serve protected embeds from
I assume this was the intention initially. Its important to serve these
embeds from a different domain than the site login cookie is set on, so
as to avoid leaking auth or other cookies to potentially untrusted embed
providers. This updates the src of the iframe appropriately to keep
cookies from being sent to it.